requirement.txt
asgiref==3.5.0
distlib==0.3.4
Django==3.2.12
djangorestframework==3.13.1
filelock==3.6.0
importlib-metadata==4.11.3
numpy==1.21.5
pandas==1.1.5
platformdirs==2.5.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
virtualenv==20.14.0
XlsxWriter==3.0.3
zipp==3.8.0